Fix moved spark imports
